Implemented connection to a server, and running is_ping.
I've ported a few things from the old code, making it no longer rely
on PHP libraries. I've also written a very thin bson layer to do
what I need, and nothing more. It's a bit of a hack, but it will need
almost no maintenance, and it's very specific; and not relying on PHP's
libraries either. All of this code in mcon should be usable from outside
the extension as well.